Add files via upload
A food aggregator company has stored the data of the different orders made by the registered customers in their online portal. They want to analyze the data to get a fair idea about the demand of different restaurants which will help them in enhancing their customer experience. The goal of the project is to perform data analysis to find answers to some key questions identified by the company that will help  to improve their business.